Initial support for creating quoted tweets vis /quote command
Adds a new /quote command to create quote tweets. After logging on the
the Twitter website (not done that in an age) to see how this is
supposed to work I see that quoted tweets just use the retweet
functionality. So... I need to rework this and try to get this
functionality under the existing /rt command as opposed to adding a
separate /quote command.

Also, found out that quote tweets are limited to 116 chars. Need to
factor that in somewhere.

References Issue #7