Move quoted tweet support under the existing /retweet command
Tested briefly and this works, but I could do with checking I've covered
all cases. The /retweet command is a bit overloaded with various options
and behaviours.

Basically the only difference between old and new (now) behaviour should
be that if text is appended to the /retweet then it is done as a quote
tweet. Old style RTs have to be called explicitly with /ort.

References Issue #7