Fix handling of display of quoted tweets in tdisplay
After commit cafec0e5c891706292d41bcb63c302968e333b55 I started seeing
an issue where streamed quoted tweets would only display the quoted
tweet, but not the parent. But on start-up or performing /again
everything *seemed* ok (i.e. parents and the quoted tweets displayed).

Streamed tweets are passed to tdisplay as an array of one item whereas
the initial call passes an array of a whole stack of tweets. So I knew
the issue was something to do with just one tweet being passed.

This led me to have a closer look at the code and I eventually realised
the importance of $disp_max in the for loop. Since if only one tweet is
passed to tdisplay it will only loop once, it was therefore only picking
up the quoted tweet and not the parent. So I now figure I was also
encountering the bug in a refresh and on start-up, but not "seeing" it
because it wasn't as obvious that tweets were being dropped.

This is yet further demonstration of me still getting to grips with
understanding the code. And Perl. And not being careful
enough.

To be safe, I'm setting the return values of tdisplay to be based upon
the original my_json_ref and not the injected version. At least until
I've figured out exactly how the return values are used.

References #6

And fingers crossed, finally fixes it!