Version 2.4.0, re-enable version checking and point at this repo on Github
It certainly would be nice to use the Github releases atom feed for this
repository to check for updates, but it is a hell of a lot easier to
just use the exact same version check file format that Floodgap use and
just point the version check url at Github instead of Floodgap.

The only downside of this is I need to remember to keep this file
updated. I *should* get around to scripting the version release process
to make this a bit more fool proof (although it'll never be me proof).

So I'm considering this as closing out this issue. I.e. it is a "good
enough" fix.

Fixes #12