[Issue 5] Allow sending long DMs
TTYtter tries to be helpful and let you know if you're sending a Tweet
that's too long, but DMs can be of allegedly unlimited lenght now, so we
should allow those to be sent.

Note that this does not change how long DMs are displayed. Long DMs are
displayed like so:

[DM dc1][FunnelFiasco/Sat Oct 31 12:14:39 +0000 2015] if all of the
raindrops were lemondrops and gumdrops oh what a rain that would be.
standing outside with my mouth
/messages/44104818-44104818

References #5