Tweak sending of newlines. Escape a \ with a \ to send literal "\n"
This is getting messy and there is no ideal solution, but the previous
commit meant that any "\n" was turned into a newline. This tweak means
that if you put "\\n" you get a literal "\n" in the tweet. Just on the
off chance someone may want to do it.

This seems like an acceptable compromise. There are still a couple of
issues though:

- This should probably still be an option so people can turn it on/off
- This means character count limitations are inaccurate. "\n" is two
  characters when composing, but only one in the actual tweet. Need to
  think about that.

References #18