TTYtter becomes oysttyer
Although mentioned elsewhere, I'll write it here as well. We got in
touch with Cameron Kaiser about making the maintenance of TTYtter a bit
more official and he said he would be willing to hand it over as long as
we changed the name, got new API keys and kept the license.

Naming stuff is hard.

- Some references correctly remain to Floodgap since that is where the
  license lies.
- Cameron Kaiser's name is retained for copyright from 2007-2012
- Some references to TTYtter remain. Similar to the references to
  Floodgap. The TERM::Readline software is still called TTYtter and so
  the version check for that still references Floodgap.