Mods to Pull Request 42, Allow users to specify oauth key/secret
Here's my thoughts on it. Pushing to a separate branch as also not sure
I should just merge in, because to be honest I don't understand the
oauth bit fully yet.

However, the `&& $pair[1] ne 'X')` mods did not work for me at all
stages of the authorisation.

What I've done "works for me" and allowed me to specify an oauth key and
secret in my `.oysttyerrc` file and obtain new tokens. But it also
allows me to switch flawlessly between them and the oysttyer oauth key
by simply commenting them out of the `.oysttyerrc` file.

The bit I'm really unsure about is removing `!length($oauthkey) ||` and
`!length($oauthsecret) ||`. It doesn't seem to have any detrimental
effect, but they must have been there for some reason in the first
place?

So, needs more investigation before we merge in.

References #42