Increase maxchars as work-around for sending long tweets with newlines
Previously, when sending newlines was introduced in
4fc104deb845c41599292d4c98c5ca05263dcff0, no attempt was made to correct
the character count. E.g:

    This is a tweet with \nnewlines\n and also a literal slash followed by n: \\n

Would be counted as 77 characters, but when sent to Twitter would only
result in 74 characters as each `\n` would only be one character and
each `\\n` would only be two characters. Therefore it is possible to
trigger oysttyer's internal maxchars (140) limit when actually the tweet
would go through fine.

The ideal thing to do would be to count things properly, but in lieu of
figuring that out we can increase the maxchars count to suit. So this is
what this commit does for the time being.

Also, add a TODO note that when replacing newlines on display of tweets
oysttyer can't yet (ever?) differentiate between real newlines and
literal slashes followed by "n"s - both will get replaced.

References #18