Fix count of characters when sending tweets with newlines.
Improve on commit 388fce17a6b64961d82b6a78eb836bdc7252a58a

Rather than increase maxchars, actually count the number of characters
to be sent correctly. I was a bit worried this might interfere with
autosplitting and truncating, but it seems to be ok, even when using
autosplit=cut/char and the 140 char limit falls on the middle of a
newline.

Add in a new length_newline function to replace the use of Perl's
standard length function. Only need to replace where relevant, not
everywhere.

Fixes #18 because I cannot be bothered to add yet another option to turn
sending of newlines on/off. If it really causes upset and people really
want to be able to write "\n" (without having to escape it as "\\n") and
it not be sent as a newline then at that point I'll do it.