Further work on destroying tco in threads
Destroying tco at the end of tdisplay does the trick everywhere. Which
got me to thinking as to whether the other calls to `destroy_all_tco`
were necessary or not. I removed a bunch of calls from `normalizejson`
and things still seem to display just fine.

If it really is ok, could also think about doing `fix_geo_api_data` in
the same place and removing the other calls to that in a similar manner.

Need to understand how and why normalizejson is used first though.

This is not ready to merge in as I don't yet understand enough the
implications of what I've done. Hence why things are just commented out
for now.

References #48