Commit to moving detco to tdisplay
I.e. delete the commented out items.

destroy_all_tco is called by normalizejson which in turn is called by
grabjson. I checked through everywhere grabjson was used to make sure
that either tdisplay was called after it or that if it wasn't then
detco-ing was not relevant.

The only one I wasn't sure about was /whois, but I've checked TTYtter
and that also doesn't detco any urls in /whois so looks like that's a
new issue (Might not be possible to detco there anyway).

But does this affect extensions?

References #48