Destroy tco in threads in normalizejson
This mostly undoes a lot of cf9c7a5388969d16760bb9ffd9ef516ef79f2932 and
puts it back to how it was before.

After much reading and following things through I realised that /thread
was calling grabjson which in turn calls normalizejson. And in normalize
json I was already detco-ing quote tweets and nested quote tweets so why
was I having to do additional detco-ing in tdisplay?

I realised that grabjson only passes through to normalizejson in certain
cases. If it is passing through an array of tweets. Or if it is a single
tweet, but also has 'liked' and 'source':

    if (defined($my_json_ref) && ref($my_json_ref) eq 'HASH' &&
        $my_json_ref->{'liked'} &&
	$my_json_ref->{'source'} &&
	((0+$my_json_ref->{'id'}) ||
	length($my_json_ref->{'id_str'}))) {

/thread would not match those conditions. I'm really not sure on the
significance of only matching on these criteria. For the time being I'm
commenting out these conditions so /thread will fall through to
normalizejson.

Yet more investigation is required before this is finished with!

References #48