Expand long DMs during start-up and /dmagain, etc (I.e. REST API)
Basically the Streaming API automatically rolled out long DM support,
but the REST API required you to specify an additonal url query
parameter.

This commit adds this in, but I don't consider it fully fixes this issue
as it has a knock-on effect. When using /dump on a dm you don't get the
full text and it is truncated in a blunt way.

I still don't understand the full background store, etc of the code, but
the way to fix this /dump location is to up all (see note below)
instances of the 1024 (byte count?) value to something that will be way
over 10000 characters. Since 1024 is suitable for 140 characters we
probably would need something like 73728. But perhaps we don't want to
show the full text in /dump anyway?

* Note: Probably not all, but that is a quick and easy way to get it to
  work instead of guessing which ones need changing and inadvertently
  breaking something.

References #36