Nicely truncate text of long dms when using /dump
Rather than display all the text of a long dm using /dump nicely
truncate it.

Use existing `csplit` subroutine to split on word boundary. However, it
required altering the `csplit` and `cosplit` routines so they expect an
`$autosplit` argument in place of the `$mode` argument. This way we can
split on the word boundary when calling `csplit` without having to set
the global `$autosplit` variable.

The only potential issue with this is whether any extensions use
`csplit`, etc. In that case it is a breaking change, but I haven't yet
found any extensions which use it.

Related to previous commit ea319c904205bf8c4a6f80dfcc79781301fe34b6

Fixes #36