Copy $EDITOR functionality for Texapp for dms
Thanks very much to @classilla for implementing this in Floodgap's
Texapp. Initially I have copied and pasted this pretty much pretty much
verbatim.

I have tried to limit the functionality to dms only. I.e so appending
`%ED%` to the end of a normal tweet will just send it. For now anyway.

Not closing this issue out yet as I haven't sorted the issue of this
always printing the message, "sorry, nothing to display."

References #57