Copy over missed code from Texapp's external editor
When I copied over the code I deleted bits I didn't think were
necessary, but because I didn't really understand it then I got a bit
over zealous and deleted the code that prefills $EDITOR with the text
typed via Oysttyer. This adds that back in.

I don't know whether to be pleased or horrified that I'm the first one
to spot my mistakes.

References #57