Move use of $EDITOR for DMs to specific /edm and /ereply commands
Just because.

I did actually think about - and have a brief play with - enabling the
%ED% approach everywhere, but it screws up with things like quote tweets
because the tweet url is appended to the end of the string screwing up
the detection of %ED%. Also, I don't really like that approach in
general of having to append something on to the end of the string when
you want to drop to an editor AND I don't like it being in updatest. It
just doesn't seem like the right place.

HOWEVER, that last bit said, not having it in updatest now screws up
newline sending. Gah! So I need to do a bit more work/thinking there.

References #57