Always prefer mp4 videos to m3u8 when both are available.
I think this is what we're after in #93. Is it the right long term
decision? Maybe. It's certainly easy to flip around if we decided to do
that at some point, but since mp4 can play natively in most places and I
had to install an extension to get m3u8 to work as a bare link in
Chrome, I think it is.